init 404 coffee project

+++ database/migrations/2026_05_24_201530_add_image_to_products_table.php
@@ -8,21 +8,29 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::table('products', function (Blueprint $table) {
+        if (!Schema::hasColumn('products', 'image')) {
 
-            $table->string('image')
-                  ->nullable()
-                  ->after('price');
+            Schema::table('products', function (Blueprint $table) {
 
-        });
+                $table->string('image')
+                      ->nullable()
+                      ->after('price');
+
+            });
+
+        }
     }
 
     public function down(): void
     {
-        Schema::table('products', function (Blueprint $table) {
+        if (Schema::hasColumn('products', 'image')) {
 
-            $table->dropColumn('image');
+            Schema::table('products', function (Blueprint $table) {
 
-        });
+                $table->dropColumn('image');
+
+            });
+
+        }
     }
 };
